refactor(repo): extract the @intake/types package (phase 3a)
Move the 29 Dexie record/data-model type declarations out of
apps/web/src/lib/db.ts into a new dependency-root package @intake/types
(packages/types/src/records.ts), consumed as raw TS via transpilePackages.

The Dexie runtime (the new Dexie() instance, the AppDatabase type, the
version().stores() migration blocks, DB_SCHEMA_VERSION and the preview
helpers) stays in db.ts, which now re-exports the record types so all ~164
@/lib/db importers resolve unchanged (zero importer churn). @intake/types is
isomorphic — no client-only/server-only/react/dexie — so it can sit at the
bottom of the dependency graph and be imported by server API routes and the
client alike.

The schema-parity extractor (dexie-schema-extractor.ts) parses the record
interfaces as text via single-file ts.createSourceFile; repoint its default
path to packages/types/src/records.ts (all 18 table interfaces kept in that
one file). Wire @intake/types into apps/web deps + next.config
transpilePackages, add packages/types/src/records.ts to the ci.yml bench
paths-filter, and update the parity-rule docs (CLAUDE.md, AGENTS.md,
@intake/db/schema comments, optional-trackers) to the new source of truth.

Phase 3a of the Turborepo migration; @intake/core (phase 3b) builds on this.

Verified: turbo typecheck + lint (0 errors), 2017 unit tests, 95
parity/journal/integrity, 33 integration (testcontainers), production build,
bundle-security scan.
